refact: SessionInfo, SessionDate, CoverImage을 SessionMakingData 클래스로 분리
